% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.stats.R
\name{boxplot.stats}
\alias{boxplot.stats}
\title{Boxplot Statistics}
\usage{
\method{boxplot}{stats}(x, coef = NULL, do.conf = TRUE, do.out = TRUE,
  statsParam = c(0.025, 0.25, 0.5, 0.75, 0.975))
}
\arguments{
\item{x}{}

\item{coef}{}

\item{do.conf}{}

\item{do.out}{}

\item{statsParam}{}
}
\description{
This function replaces boxplot.stats --- it uses the middle 50% and middle 90% instead of middle 50%(IQR) and 1.5*IQR.
}
\examples{

#Boxplot Statistics

}
\keyword{Boxplot}
\keyword{Statistics}
