% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGibbsSumamry.R
\name{getGibbsSummary}
\alias{getGibbsSummary}
\title{Summarize Output of the CDM Model}
\usage{
getGibbsSummary(ssOut, burnin = NULL, colNames = NULL, sigmaPerSeason = T)
}
\arguments{
\item{ssOut}{CDM output list.}

\item{burnin}{Number of burning .}

\item{colNames}{Name of each variable in the output.}

\item{sigmaPerSeason}{whether each site/season has a separate process error}
}
\description{
This function return a summary of the output from the Gibbs-Sampling of the CDM model.
}
\examples{

#Summarize CDM Model Ouput
ssSim <- phenoSim(nSites = 4, #number of sites
                  nTSet = 20, #number of Time steps
                  beta = c(1, 2), #beta coefficients
                  sig = .01, #process error
                  tau = .2, #observation error
                  plotFlag = T, #whether plot the data or not
                  miss = 0.1, #fraction of missing data
                  ymax = c(9,5,7, 3) #maximum of saturation trajectory
)

ssOut <- fitCDM(x = ssSim$x, #predictors  
                nGibbs = 2000,
                nBurnin = 1000,
                z = ssSim$z,#response
                connect = ssSim$connect, #connectivity of time data
                quiet=T)

summ <- getGibbsSummary(ssOut, burnin = 1000, sigmaPerSeason = F)

colMeans(summ$ymax)
colMeans(summ$betas)
colMeans(summ$tau)
colMeans(summ$sigma)

}
\keyword{Gibbs}
\keyword{Output}
\keyword{Sammary}
\keyword{Sampling}
