% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{plotPost}
\alias{plotPost}
\title{Plot Posterior Distributions}
\usage{
plotPost(chains, trueValues = NULL, outline = F)
}
\arguments{
\item{chains}{Gibbs sampling chains}

\item{trueValues}{numeric vector of true values}
}
\description{
This function plot posterior distributions of the parameters.
}
\examples{

#Summarize CDM Model Ouput

}
\keyword{Distributions}
\keyword{Plot}
\keyword{Posterior}
