% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoSimPlot.R
\name{phenoSimPlot}
\alias{phenoSimPlot}
\title{Plot Simulated Phenology Data}
\usage{
phenoSimPlot(z, connect, add = F, col = "blue", ylim = range(z, na.rm =
  T), pch = 1, lwd = 1)
}
\arguments{
\item{z}{A vector of time-series data [n x 1]}

\item{connect}{The connectivity matrix for the z vector [n x 2]. Each row contains the last and next elements of the time-series. NA values means not connected.}

\item{add}{A logical variable whether the plot should be overlaid on the current panel.}

\item{col}{The color variable}

\item{ylim}{Range of the y axis}

\item{pch}{pch value for the points}

\item{lwd}{lwd value for line tickness}
}
\description{
This function plots the time-series data described with a connectivity matrix.
}
\examples{

#Plot Simulated Phenology Data
ssSim <- phenoSim(nSites = 4, #number of sites
                  nTSet = 20, #number of Time steps
                  beta = c(1, 2), #beta coefficients
                  sig = .01, #process error
                  tau = .2, #observation error
                  plotFlag = T, #whether plot the data or not
                  miss = 0.1, #fraction of missing data
                  ymax = c(9,5,7, 3) #maximum of saturation trajectory
)

ssOut <- fitCDM(x = ssSim$x, #predictors
                nGibbs = 2000,
                nBurnin = 1000,
                z = ssSim$z,#response
                connect = ssSim$connect, #connectivity of time data
                quiet=T)

summ <- getGibbsSummary(ssOut, burnin = 1000, sigmaPerSeason = F)

colMeans(summ$ymax)
colMeans(summ$betas)
colMeans(summ$tau)
colMeans(summ$sigma)

}
\keyword{Data}
\keyword{Phenology}
\keyword{Plot}
\keyword{Simulated}
