% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPOGibbs.R
\name{plotPOGibbs}
\alias{plotPOGibbs}
\title{Plot Observed vs Predicted}
\usage{
plotPOGibbs(o, p, nburnin = NULL, xlim = range(o, na.rm = T),
  ylim = range(p, na.rm = T), xlab = "Observed", ylab = "Predicted",
  colSet = c("#fb8072", "#80b1d3", "black"), cex = 1, lwd = 2, pch = 19)
}
\arguments{
\item{o}{Observed vector}

\item{p}{Predicted Gibbs samples}
}
\description{
This function plot posterior distributions of the parameters.
}
\examples{

#Plot Observed vs Predicted

}
\keyword{Observed}
\keyword{Plot}
\keyword{Predicted}
\keyword{vs}
